removed resolveBucketLink
